idShiftUtils: separate ID-map functions into dedicated package.
